+++ packages/vue/CHANGELOG.md
@@ -8,6 +8,10 @@ description: All notable changes to this project will be documented in this file
 
 ### Fixed
 
+- Resolved an issue whre the `Toast` component would not render custom content.
+
+### Fixed
+
 - Resolved an issue where the `SegmentGroup` component would not animate on the first click.
 - Fixed an issue where standalone imports were not working as expected.